+++ test_oai_pmh_extractor/test_payload_collector.py
@@ -20,10 +20,12 @@ from oai_pmh_extractor import (
     [
         (TransferUrlFilters.filter_by_regex("([a-z]+)"), ["abc", "d", "ef"]),
         (TransferUrlFilters.filter_by_regex("([0-9]+)"), ["123", "456"]),
+        (TransferUrlFilters.filter_by_regex(".*"), ["123abc456d ef"]),
     ],
     ids=[
         "letters",
         "digits",
+        "match anything"
     ]
 )
 def test_filter_by_regex(this_filter, expected_result):
@@ -57,10 +59,246 @@ def test_filter_by_regex(this_filter, expected_result):
 )
 def test_filter_by_regex_in_xml_path(source_metadata, expected_result):
     """Test factory `filter_by_regex_in_xml_path` of `TransferUrlFilters`."""
-    this_filter = \
-        TransferUrlFilters.filter_by_regex_in_xml_path("([a-z]+)", ["root", "a"])
+    this_filter = TransferUrlFilters.filter_by_regex_in_xml_path(
+        "([a-z]+)", ["root", "a"]
+    )
 
     assert expected_result == this_filter(source_metadata)
+
+
+def test_filter_by_regex_in_xml_path_match_anything():
+    """
+    Test factory `filter_by_regex_in_xml_path` of `TransferUrlFilters`
+    for a regex that matches anything.
+
+    Assert empty matches are removed.
+    """
+    this_filter = TransferUrlFilters.filter_by_regex_in_xml_path(
+        ".*", ["root", "a"]
+    )
+
+    assert ["asd"] == this_filter("<root><a>asd</a></root>")
+
+
+@pytest.mark.parametrize(
+    ("source_metadata", "input_regex", "input_xpath", "expected_result"),
+    [
+        (
+            "<root><a>123abc123</a><a>123def123</a></root>",
+            "([a-z]+)",
+            "./a",
+            ["abc", "def"],
+        ),
+        (
+            "<root><a>123abc123def123</a></root>",
+            "([a-z]+)",
+            "./a",
+            ["abc", "def"],
+        ),
+        (
+            "<root><a>123abc123def123</a><b>123abc123</b></root>",
+            "([a-z]+)",
+            "./a",
+            ["abc", "def"],
+        ),
+        (
+            "<root><a>123abc123</a><a/><a>123def123</a></root>",
+            "([a-z]+)",
+            "./a",
+            ["abc", "def"],
+        ),
+        ("<root><a>123</a></root>", "([a-z]+)", "./a", []),
+        ("<root><a/></root>", "([a-z]+)", "./a", []),
+        ("<root><a>123abc123</a><a/></root>", "([a-z]+)", "./a", ["abc"]),
+        ("<root><b>123</b></root>", "([a-z]+)", "./a", []),
+        (
+            "<root><a>123abc123</a><a>123def123</a></root>",
+            "([a-z]+)",
+            "./a[1]",
+            ["abc"],
+        ),
+        (
+            "<root><a>123abc123</a><a>123def123</a></root>",
+            ".*",
+            "./a",
+            ["123abc123", "123def123"],
+        ),
+    ],
+    ids=[
+        "multiple tags with matches",
+        "tag with multiple matches",
+        "matching contents at different path",
+        "multiple tags with matches and an empty tag",
+        "no match",
+        "single empty tag",
+        "match and empty tag",
+        "xpath non-existent",
+        "multiple tags with matches - select first",
+        "multiple tags with matches - match anything",
+    ],
+)
+def test_filter_by_regex_with_xpath_query(
+    source_metadata, input_regex, input_xpath, expected_result
+):
+    """Test factory `filter_by_regex_with_xpath_query` of `TransferUrlFilters`."""
+    this_filter = TransferUrlFilters.filter_by_regex_with_xpath_query(
+        input_regex, input_xpath
+    )
+
+    assert expected_result == this_filter(source_metadata)
+
+
+@pytest.mark.parametrize(
+    ("source_metadata", "input_regex", "input_xpath", "expected_result"),
+    [
+        (
+            """<?xml version="1.0" encoding="UTF-8"?>
+<root xmlns="uri_xmlns/">
+  <namespace_a:dc xmlns:namespace_a="uri_xmlns/namespace_a/" xmlns:dc="uri_dc">
+    <dc:identifier>123abc123</dc:identifier>
+    <dc:identifier>123123</dc:identifier>
+  </namespace_a:dc>
+</root>""",
+            "([a-z]+)",
+            "./namespace_a:dc/dc:identifier",
+            ["abc"],
+        ),
+        (
+            """<?xml version="1.0" encoding="UTF-8"?>
+<root xmlns="uri_xmlns/">
+  <namespace_a:dc xmlns:namespace_a="uri_xmlns/namespace_a/" xmlns:dc="uri_dc">
+    <dc:identifier>123abc123</dc:identifier>
+    <dc:identifier>123def123</dc:identifier>
+  </namespace_a:dc>
+</root>""",
+            "([a-z]+)",
+            "./namespace_a:dc/dc:identifier",
+            ["abc", "def"],
+        ),
+        (
+            """<?xml version="1.0" encoding="UTF-8"?>
+<root xmlns="uri_xmlns/">
+  <namespace_a:dc xmlns:namespace_a="uri_xmlns/namespace_a/" xmlns:dc="uri_dc">
+    <dc:identifier>123123</dc:identifier>
+  </namespace_a:dc>
+</root>""",
+            "([a-z]+)",
+            "./namespace_a:dc/dc:identifier",
+            [],
+        ),
+        (
+            """<?xml version="1.0" encoding="UTF-8"?>
+<root xmlns="uri_xmlns/">
+  <namespace_a:dc xmlns:namespace_a="uri_xmlns/namespace_a/" xmlns:dc="uri_dc">
+    <dc:identifier>123abc123</dc:identifier>
+    <dc:identifier/>
+  </namespace_a:dc>
+</root>""",
+            "([a-z]+)",
+            "./namespace_a:dc/dc:identifier",
+            ["abc"],
+        ),
+        (
+            """<?xml version="1.0" encoding="UTF-8"?>
+<root xmlns="uri_xmlns/">
+  <namespace_a:dc xmlns:namespace_a="uri_xmlns/namespace_a/" xmlns:dc="uri_dc">
+    <dc:identifier>123abc123</dc:identifier>
+  </namespace_a:dc>
+  <namespace_b:dc xmlns:namespace_b="uri_xmlns/namespace_b/" xmlns:dc="uri_dc">
+    <dc:identifier>123def123</dc:identifier>
+  </namespace_b:dc>
+</root>""",
+            "([a-z]+)",
+            "./namespace_b:dc/dc:identifier",
+            ["def"],
+        ),
+    ],
+    ids=[
+        "single match",
+        "multiple matches",
+        "no match",
+        "with empty tag",
+        "several fields at different locations",
+    ],
+)
+def test_filter_by_regex_with_xpath_query_realistic_metadata(
+    source_metadata, input_regex, input_xpath, expected_result
+):
+    """
+    Test factory `filter_by_regex_with_xpath_query` of `TransferUrlFilters`.
+    """
+    this_filter = TransferUrlFilters.filter_by_regex_with_xpath_query(
+        input_regex, input_xpath
+    )
+
+    assert expected_result == this_filter(source_metadata)
+
+
+def test_filter_by_regex_with_xpath_query_non_existent_namespace():
+    """
+    Test factory `filter_by_regex_with_xpath_query` of `TransferUrlFilters`.
+    """
+    this_filter = TransferUrlFilters.filter_by_regex_with_xpath_query(
+        "", "./namespace_b:dc/dc:identifier"
+    )
+
+    with pytest.raises(SyntaxError) as exc_info:
+        this_filter(
+            """<?xml version="1.0" encoding="UTF-8"?>
+<root xmlns="uri_xmlns/">
+  <namespace_a:dc xmlns:namespace_a="uri_xmlns/namespace_a/" xmlns:dc="uri_dc">
+    <dc:identifier>123abc123</dc:identifier>
+  </namespace_a:dc>
+</root>"""
+        )
+    assert "prefix 'namespace_b' not found in prefix map" in str(
+        exc_info.value
+    )
+
+
+def test_download_record_payload_for_syntaxerror_in_transfer_url_filter():
+    """
+    Test method `known download_record_payload` of PayloadCollector class
+    for a known SyntaxError from `transfer_url_filter`.
+    """
+
+    msg = "prefix 'namespace_b' not found in prefix map"
+    def fake_transfer_url_filter(*args):
+        raise SyntaxError(msg)
+
+    some_payload_collector = PayloadCollector(transfer_url_filters=[fake_transfer_url_filter])
+    with mock.patch.object(some_payload_collector, "download_file"):
+        # test single url
+        record = OAIPMHRecord("")
+        some_payload_collector.download_record_payload(
+            record=record,
+            path=TEST_DIRECTORY,
+            skip_download=True
+        )
+    assert msg in some_payload_collector.log[Context.ERROR][0].body
+
+
+def test_download_record_payload_for_exception_in_transfer_url_filter():
+    """
+    Test method `download_record_payload` of PayloadCollector class
+    for an unknown exception from `transfer_url_filter`.
+    """
+
+    msg = "error message"
+    def fake_transfer_url_filter(*args):
+        raise Exception(msg)
+
+    some_payload_collector = PayloadCollector(transfer_url_filters=[fake_transfer_url_filter])
+    with mock.patch.object(some_payload_collector, "download_file"):
+        # test single url
+        record = OAIPMHRecord("")
+        with pytest.raises(Exception) as exc_info:
+            some_payload_collector.download_record_payload(
+                record=record,
+                path=TEST_DIRECTORY,
+                skip_download=True
+            )
+    assert msg == str(exc_info.value)
 
 
 SIMPLE_TRANSFER_URLS = [
@@ -83,6 +321,38 @@ def simple_payload_collector():
     # build PayloadCollector with faked mapper
     return PayloadCollector(
         transfer_url_filter
+    )
+
+
+def test_payload_collector_constructor():
+    """
+    Test constructor of `PayloadCollector` class for the mutually exclusive
+    kwargs 'transfer_url_filters' and 'transfer_url_filter'.
+    """
+    with pytest.raises(TypeError) as exc_info:
+        PayloadCollector(
+            transfer_url_filters=[lambda x: []],
+            transfer_url_filter=lambda x: []
+        )
+    assert str(exc_info.value) == (
+        "Cannot instantiate 'PayloadCollector' with given args. "
+        + "Exactly one of the kwargs 'transfer_url_filters' and "
+        + "'transfer_url_filter' has to be specified."
+    )
+
+    with pytest.raises(TypeError) as exc_info:
+        PayloadCollector()
+    assert str(exc_info.value) == (
+        "Cannot instantiate 'PayloadCollector' with given args. "
+        + "Exactly one of the kwargs 'transfer_url_filters' and "
+        + "'transfer_url_filter' has to be specified."
+    )
+
+    assert isinstance(
+        PayloadCollector(transfer_url_filters=[lambda x: []]), PayloadCollector
+    )
+    assert isinstance(
+        PayloadCollector(transfer_url_filter=lambda x: []), PayloadCollector
     )
 
 
@@ -139,6 +409,7 @@ def test_download_file_filename_from_header(
 
         assert used_filename.name == test_filename
 
+
 def test_download_file_filename_from_url(
     simple_payload_collector
 ):
@@ -192,6 +463,7 @@ def test_download_file_filename_from_url(
 
         assert used_filename.name in test_url
 
+
 def test_download_file_filename_failing(
     simple_payload_collector
 ):
@@ -244,6 +516,7 @@ def test_download_file_filename_failing(
             simple_payload_collector.download_file(TEST_DIRECTORY, test_url)
 
     assert test_url in simple_payload_collector.log[Context.ERROR][0].body
+
 
 def test_download_file_filename_override(
     simple_payload_collector
@@ -362,8 +635,9 @@ def test_download_file_faked(simple_payload_collector):
         written_content = called_path_class.write_bytes.call_args_list[0][0][0]
         assert written_content == MockedResponse(SIMPLE_TRANSFER_URLS[0]).read()
 
+
 # switch these two lines to in-/exclude this test in suite
-#def notest_download_file_encoding(simple_payload_collector):
+# def notest_download_file_encoding(simple_payload_collector):
 def test_download_file_encoding(simple_payload_collector):
     """
     Actually test method download_file of PayloadCollector class.
@@ -385,6 +659,7 @@ def test_download_file_encoding(simple_payload_collector):
     # post-test cleanup
     if TEST_DIRECTORY.is_dir():
         shutil.rmtree(TEST_DIRECTORY)
+
 
 def test_download_record_payload_success():
     """
@@ -425,6 +700,139 @@ def test_download_record_payload_success():
             assert file["complete"]
         assert some_payload_collector.download_file.call_count == \
             len(MORE_TRANSFER_URLS)
+
+
+def test_download_record_payload_duplicate_urls():
+    """
+    Test method download_record_payload of PayloadCollector class
+    for duplicate urls.
+
+    To this end, patch the download_file-method to do nothing. Evaluate
+    number of calls afterwards.
+    """
+
+    TRANSFER_URLS_1 = [
+        "1",
+        "2",
+        "1"
+    ]
+    TRANSFER_URLS_2 = [
+        "3",
+        "3",
+        "2"
+    ]
+
+    some_payload_collector = PayloadCollector(
+        transfer_url_filters=[
+            lambda x: TRANSFER_URLS_1,
+            lambda x: TRANSFER_URLS_2
+        ]
+    )
+    with mock.patch.object(some_payload_collector, "download_file"):
+        record = OAIPMHRecord("")
+        some_payload_collector.download_record_payload(
+            record=record,
+            path=TEST_DIRECTORY
+        )
+        for file in record.files:
+            assert file["complete"]
+        assert some_payload_collector.download_file.call_count == 3
+
+
+def test_download_record_payload_renew_urls():
+    """
+    Test method download_record_payload of PayloadCollector class
+    with renew_urls.
+
+    To this end, patch the download_file-method to do nothing. Evaluate
+    number of calls afterwards.
+    """
+
+    SINGLE_TRANSFER_URLS = [
+        "a"
+    ]
+    MORE_TRANSFER_URLS = [
+        "1",
+        "2"
+    ]
+
+    some_payload_collector = PayloadCollector(
+        transfer_url_filters=[
+            lambda x: SINGLE_TRANSFER_URLS,
+            lambda x: MORE_TRANSFER_URLS,
+        ]
+    )
+    with mock.patch.object(some_payload_collector, "download_file"):
+        # test single url
+        record = OAIPMHRecord("", file_urls=SINGLE_TRANSFER_URLS)
+        some_payload_collector.download_record_payload(
+            record=record,
+            path=TEST_DIRECTORY,
+            renew_urls=True
+        )
+        assert some_payload_collector.download_file.call_count == \
+            len(SINGLE_TRANSFER_URLS) + len(MORE_TRANSFER_URLS)
+        assert record.files[0]["complete"]
+        # reset
+        some_payload_collector.download_file.reset_mock()
+        # assert multiple urls
+        record = OAIPMHRecord("", file_urls=MORE_TRANSFER_URLS)
+        some_payload_collector.download_record_payload(
+            record=record,
+            path=TEST_DIRECTORY
+        )
+        for file in record.files:
+            assert file["complete"]
+        assert some_payload_collector.download_file.call_count == \
+            len(MORE_TRANSFER_URLS)
+
+
+def test_download_record_payload_no_files():
+    """
+    Test method download_record_payload of PayloadCollector class
+    for record with an empty list/None in record.files.
+
+    To this end, patch the download_file-method to do nothing. Evaluate
+    number of calls afterwards.
+    """
+
+    SINGLE_TRANSFER_URLS = [
+        "a"
+    ]
+    MORE_TRANSFER_URLS = [
+        "1",
+        "2"
+    ]
+
+    some_payload_collector = PayloadCollector(
+        transfer_url_filters=[
+            lambda x: SINGLE_TRANSFER_URLS,
+            lambda x: MORE_TRANSFER_URLS,
+        ]
+    )
+    with mock.patch.object(some_payload_collector, "download_file"):
+        # test single url
+        record = OAIPMHRecord("", file_urls=[])
+        some_payload_collector.download_record_payload(
+            record=record,
+            path=TEST_DIRECTORY
+        )
+        assert some_payload_collector.download_file.call_count == \
+            len(SINGLE_TRANSFER_URLS) + len(MORE_TRANSFER_URLS)
+        assert record.files[0]["complete"]
+        # reset
+        some_payload_collector.download_file.reset_mock()
+        # assert multiple urls
+        record = OAIPMHRecord("", file_urls=MORE_TRANSFER_URLS)
+        some_payload_collector.download_record_payload(
+            record=record,
+            path=TEST_DIRECTORY
+        )
+        for file in record.files:
+            assert file["complete"]
+        assert some_payload_collector.download_file.call_count == \
+            len(MORE_TRANSFER_URLS)
+
 
 def test_download_files_partial_fail():
     """
@@ -471,6 +879,7 @@ def test_download_files_partial_fail():
         assert some_payload_collector.download_file.call_count == \
             len(MORE_TRANSFER_URLS)
         assert error_msg in some_payload_collector.log[Context.ERROR][0].body
+
 
 def test_download_file_file_url(simple_payload_collector):
     """